--- src/App.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import './App.css';
-import './Hero';
-import Hero from './Hero';
-
-class App extends Component {
-    render() {
-        return (
-            <Hero />
-        );
-    }
-}
-
-export default App;
