+++ src/App/index.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import Hero from "../Hero/index.js";
+
+class App extends Component {
+  render() {
+    return <Hero />;
+  }
+}
+
+export default App;
