--- src/Hero.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import './Hero.css';
-
-class Hero extends Component {
-    render() {
-        return (
-            <div id='hero'>
-                <div className='card'>
-                    <h1>Amélégumes</h1>
-                </div>
-            </div>
-        );
-    }
-}
-
-export default Hero;
