+++ src/Hero/index.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import "./style.css";
+
+class Hero extends Component {
+  render() {
+    return (
+      <div id="hero">
+        <div className="header">
+          <h1>Amélégumes</h1>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Hero;
