+++ src/Hero/style.css
@@ -1,0 +1,34 @@
+@font-face {
+  font-family: "Pacifico";
+  src: local("Pacifico"), url(Pacifico-Regular.ttf) format("truetype");
+}
+
+#hero {
+  align-items: center;
+  background: url("background.jpg") no-repeat center center fixed;
+  background-size: cover;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  min-height: 100vh;
+
+  /* https://stackoverflow.com/questions/51795887/how-to-make-body-take-100-min-height-and-height-as-auto/51796359 */
+  /* overflow: auto; */
+}
+
+.header {
+  background-color: rgba(255, 255, 255, 0.5);
+  border-radius: 1em;
+  padding: 1em;
+}
+
+h1 {
+  font-family: "Pacifico", cursive;
+  font-size: 3em;
+}
+
+@media only screen and (min-width: 500px) {
+  h1 {
+    font-size: 5em;
+  }
+}
